backend/requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pymongo==3.12.0
black==21.12b0
bcrypt==3.2.0
pydantic==1.10.0
pydantic_core==2.4.0
PyJWT==2.8.0
python-decouple==3.5
